third_party/blink/web_tests/fast/dom/StyleSheet/detached-style-expected.txt

Bug 45205: Test that <link> elements only have stylesheets when in document.

Testing <link>...
PASS sheet.ownerNode is link
PASS link.sheet === sheet is true
Removing <link>...
PASS link.sheet is null
PASS sheet.ownerNode is null
PASS sheet.cssRules[0].cssText[0] is "*"
Re-adding <link>...
PASS sheet.ownerNode is null
PASS link.sheet === sheet is false
PASS link.sheet.ownerNode is link
Testing <style>...
PASS sheet.ownerNode is style
PASS style.sheet is sheet
Removing <style>...
PASS style.sheet is null
PASS sheet.ownerNode is null
PASS sheet.cssRules[0].cssText[0] is "*"
Re-adding <style>...
PASS sheet.ownerNode is null
PASS style.sheet === sheet is false
PASS style.sheet.ownerNode is style
PASS successfullyParsed is true

TEST COMPLETE
